+++ CHANGELOG.md
@@ -6,6 +6,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [PEP 440](https://www.python.org/dev/peps/pep-0440/)
 and uses [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.3.2]
+
+### Changed
+- Reference and secondary scenes are now accepted in any order, but must still have the same set of burst number + swath + polarization identifiers.
+
+### Fixed
+- Reference scenes are now rejected if they contain any duplicate burst number + swath + polarization identifiers.
+- Allow scenes over water when water masking is applied. Fixes https://github.com/ASFHyP3/hyp3-isce2/issues/211 and https://github.com/ASFHyP3/hyp3-isce2/issues/247.
+- Pinned to `burst2safe>=1.4.5` to address https://github.com/ASFHyP3/burst2safe/issues/170.
+
 ## [2.3.1]
 
 ### Changed